+++ Changelog.md
@@ -7,8 +7,9 @@ Language Features:
 
 
 Compiler Features:
+ * Commandline Interface: Add ``--include-path`` option for specifying extra directories that may contain importable code (e.g. packaged third-party libraries).
  * Commandline Interface: Do not implicitly run evm bytecode generation unless needed for the requested output.
- * Commandline Interface: Normalize paths specified on the command line and make them relative for files located inside base path.
+ * Commandline Interface: Normalize paths specified on the command line and make them relative for files located inside base path and/or include paths.
  * Immutable variables can be read at construction time once they are initialized.
  * SMTChecker: Add constraints to better correlate ``address(this).balance`` and ``msg.value``.
  * SMTChecker: Support constants via modules.